refactor: razorpay functions are moved to  folder
