+++ src/Routes/Cart/Cart.js
@@ -1,85 +1,18 @@
 import { useContext, useEffect, useState } from "react";
-import { getCartItems, verifyPayment } from "../../utils/helper";
+import { getCartItems, calculateOrderPrice } from "../../utils/helper";
+import { displayRazorpay } from "../../utils/razorpay";
 import AuthContext from '../../store/auth-context';
-import { calculateOrderPrice, placeOrder } from "../../utils/helper";
-
-const RZP_KEY_ID = 'rzp_test_lAIUDkdEDcflFM';
 
 
 function PlaceOrder(props) {
-    // const [orderId, setOrderId] = useState(null);
-
-    function loadScript(src) {
-        return new Promise((resolve) => {
-            const script = document.createElement("script");
-            script.src = src;
-            script.onload = () => {
-                resolve(true);
-            };
-            script.onerror = () => {
-                resolve(false);
-            };
-            document.body.appendChild(script);
-        });
-    }
-
-    const displayRazorpay = async () => {
-        const res = await loadScript(
-            "https://checkout.razorpay.com/v1/checkout.js"
-        );
-
-        if (!res) {
-            alert("Razorpay SDK failed to load. Are you online?");
-            return;
-        }
-
-        // creating a new order
-        const order = await placeOrder(props.orderPrice, props.token);
-
-        console.log(order);
-        if (order.error) {
-            alert("Server error. Are you online?");
-            return;
-        }
-
-        // getting order details
-        const {id, amount, currency} = order.data;
-
-        const options = {
-            key: RZP_KEY_ID, 
-            amount: amount.toString(),
-            currency: currency,
-            name: "Mount Pizza",
-            description: "Test Transaction",
-            order_id: id,
-            handler: async function (response) {
-                const data = {
-                    orderCreationId: id,
-                    razorpayPaymentId: response.razorpay_payment_id,
-                    razorpayOrderId: response.razorpay_order_id,
-                    razorpaySignature: response.razorpay_signature,
-                };
-                const paymentStatus = await verifyPayment(props.token, data);
-                console.log(paymentStatus);
-            },
-            prefill: {
-                name: "Karl Urban",
-                email: "karl@gmail.com",
-            },
-            theme: {
-                color: "#61dafb",
-            },
-        };
-
-        // creating razorpay payment window
-        const paymentObject = new window.Razorpay(options);
-        paymentObject.open();
+    const handleOrderPlacement = async () => {
+        displayRazorpay(props);
     }
 
     return (
         <>
             {
-                <button onClick={displayRazorpay}>
+                <button onClick={handleOrderPlacement}>
                     {`Pay ${props.orderPrice}`}
                 </button>
             }
@@ -90,6 +23,7 @@ function PlaceOrder(props) {
 export default function Cart() {
     const [items, setItems] = useState([]);
     const [price, setPrice] = useState(0);
+    const [isOrderPlaced, setIsOrderPlaced] = useState(false);
     const authCtx = useContext(AuthContext);
 
     useEffect(() => {
@@ -108,6 +42,8 @@ export default function Cart() {
     useEffect(() => {
         setPrice(calculateOrderPrice(items));
     }, [items])
+
+    useEffect(() => {}, [isOrderPlaced])
 
     return (
         <>
@@ -131,13 +67,25 @@ export default function Cart() {
                 })
             }
 
-            <div>
-                <strong>Order Price: </strong> {price}
-            </div>
+            {
+                items.length > 0
+                ?
+                <>
+                    <div>
+                        <strong>Order Price: </strong> {price}
+                    </div>
 
-            <div>
-                <PlaceOrder token={authCtx.token} orderPrice={price} />
-            </div>
+                    <div>
+                        <PlaceOrder token={authCtx.token} setIsOrderPlaced={setIsOrderPlaced} orderPrice={price} />
+                    </div>
+                </>
+                :
+                <>
+                    <div>
+                        <strong>CART IS EMPTY</strong>
+                    </div>
+                </>
+            }
         </>
     );
 }
